Split strategies into separate modules
