+++ aquadoggo/src/db/models/mod.rs
@@ -10,7 +10,7 @@ mod query;
 mod task;
 pub mod utils;
 
-pub use self::log::{LogRow, LogHeightRow};
+pub use self::log::{LogHeightRow, LogRow};
 pub use document::{DocumentRow, DocumentViewFieldRow};
 pub use entry::EntryRow;
 pub use operation::{OperationFieldsJoinedRow, OperationRow};
